requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
openai>=1.40
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
